Clean up schema_reader.py a bit
